requirements.txt
# Core data science and numerical computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Financial data and analysis
yfinance>=0.2.18
quantlib>=1.31

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Statistical testing
statsmodels>=0.14.0

# Development and testing
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.23.0
nbconvert>=7.4.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional: Advanced optimization (for potential future enhancements)
cvxpy>=1.3.0